Use gopsutil to search by process full pattern on Windows
